# Add initial performance tests for XmlConvert.