@TrayanZapryanov 2022-02-07
I am sorry but I was not able to check against 7.0. I've tried to follow guides, but couldn't find how to build.
I am receiving this error: >C:\Program Files\dotnet\sdk\6.0.101\Sdks\Microsoft.NET.Sdk\targets\Microsoft.NET.TargetFrameworkInference.targets(144,5): error NETSDK1045: The current .NET SDK does not support targeting .NET 7.0.  Either target .NET 6.0 or lower, or use a version of the .NET SDK that supports .NET 7.0.