@TrayanZapryanov 2022-02-07
> @TrayanZapryanov no worries, it's a little unusual. Can you install the latest nightly 7.0 SDK from https://github.com/dotnet/installer?
> 
> https://github.com/dotnet/performance/blob/b73570d6eea0e85003ae0b6d018c5c46eb7daef0/src/benchmarks/micro/README.md should link to this https://github.com/dotnet/performance/blob/b73570d6eea0e85003ae0b6d018c5c46eb7daef0/docs/prerequisites.md#net-core-sdk
> 
> I will clarify the docs as soon as we figure out where you tripped up.

Thanks for the hint - only installer was missing.
I was able to build against my Debug build and soon will test against Release too.