@TrayanZapryanov 2022-02-07
One more thing that I was not able to solve is how to force runtime build to create Release ""testhost"" folder.
I've tried :
build.cmd -rc Release
build.cmd clr+libs -rc Release

but always found net7.0-windows-Debug-x64 after build