@TrayanZapryanov 2022-02-07
One last thing that I cannot understand is why new results are slower : they have better memory profile....
Also I used following command to build and run benchmarks:
dotnet run -c Release -f net7.0 --filter *XmlConvert* --coreRun C:\Work\Github\TZ_runtime\artifacts\bin\testhost\net7.0-windows-Release-x64\shared\Microsoft.NETCore.App\7.0.0\CoreRun.exe